backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
psycopg2-binary==2.9.9
requests==2.31.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
